+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Fixed
 
 - Fix align on fake-split charts
-- Fix fake-split charts intermediate step jump.
+- Drilldown on split chart is fade.
+- Fix aggregate on split chart.
 - Do not interpolate hiding/showing legend
 - Fix aggregator interface for 'set' channel parameter:
   - From now not accepted the same dimension on the same channel.